Add files via upload
PlayerFinal is the main player control script, make sure there is a rigidbody and tag set to "Player". Follow Cam needs to be on the Main Camera, the Target is the player plane. The Rig is the "Rig" child from the player plane prefab. Trail Follow goes on an empty game object.